Wrote notes, made tables, and entered data
This is the beginnings of the project. The proposals Markdown are
general notes, the openrpg.sql shows the table structure for the
beginning tables, and characters.csv shows the preliminary data
(for characters).

Observations as I make work on this project? I'd really like to
learn how to set up foreign key constraints while working on tables.
I'd also like a few more options on importing data to tables. I'd
like to try using a GUI tool every once in a while, like phpmyadmin.