Index.php Produces for filling out events
In this example, the game master can take a handful of characters
and use the form to write the "events" in that characters time line.
This time line would be simliar to a facebook style time line.

So far, the index.php gets the names from the database, and puts
info necessary for a form (to be done via ajax) into a jsonp file
names event.jsonp. I'm not going to track the file in git as it
will be frenquently overwritten.