Changed MySQL tables
I realized in the killings table I made I did not properly reference
death_id's correctly. Specifically, the killings table only included
entries for character_ids, but I labeled them as death_ids. Now there
is a new column that properly references the death_id.

I also changed victims to victim_id so as to get be clear as to which
columns will join on which tables