Removed_from_play field added to deaths table
The possibility arose that in an rpg setting. You might want to
record a character death without the character being considered
removed from play. Character might get resurrected, be undead,
survive as a ghost, etc. and thus be playable after having died.

The code doesn't support this feature, and table constraints (primary
keys) need to be dropped as well for the computer to not encounter an
error.

It's also worth mentioning that a transaction, once started, needs to
be committed. Otherwise it will never record. I've added the pdo::commit
method here.