Added Death Codes functionality
In my management system, I've added the idea of death codes. In
this case, you can specify precisely the kind of "death" that occurs
because in RPG's "death" can be somewhat ambiguous. A death code
thus refers to whether you've died a biological death, been turned
into various forms of undead, be turned to stone, etc.

The death codes are as follows: Death code starts at 0 (you no longer
exist), Add 1 if the body is still present (death). Add 2 if you
can move. Add 4 if your normal biology is the same. Add 8 if
free will remains. Add 16 if the character still retains memory of
previous existence. Add 32 if the condition can be reversed
by the same means the "death" was caused, such as a spell that
can be dissipated.

Scores have been calculated for common RPG deaths: annihilation,
death, petrification, zombification, etc. They are in a death_codes
SQL Table. Plans are also to add 64 (or more) to create
a custom code. This would be if multiple types of death events
that need to be distinguished for "death types" should they
end up with the same score (i.e. the six questions distinguishing
death events fail to make a necessary distinction. A test case
might be zombies. Zombies from viruses, zombies from parasites,
and supernatural zombies all mindless crave the flesh and brains
of humans, but an RPG based on Resident Evil (where multiple
zombie types coexist). Might need to track those differences.