Wrote page to select a card at random
I've written a page that will select a candidate card at random.
One is a static page for easy viewing, and one is a dynamic page
for easy updating. Both will show a random candidate card, but
the static page has some CSS tweaks not yet shared with the dynamic