beginning hand project for superpac
My new project, in cardhand.php is to pull a group of cards that
would be in superpac into a hand full of cards. From there it's time
to play them.

A thing left undone in my other project, openrpg, is using the jquery
plugin chosen to get better select menu buttons. Currently this is
not functioning.

Next step is to start modifying the css file I'm using as a template