Initial: Super Pac Internals: Having a Hand
In order for my Super Pac program to run, the script has to keep
track of game information and display it. Currently the script holds
a hand of action cards, which can be added to with the draw button,
displayed with the show button, and reset with the reset button.
This information currently only shows up in the console. How to display
it is another matter.