New Files: Super PAC UI as wrappers around jquery
I've experienced some difficulty in creating a working UI for Super
Pac. Of great concern is the difficulty in meshing the jquery
representation of an object along with it's associated data. In this
approach I create a DOM object with jquery that has the data it needs
attached. This way I can fiddle how it should look without having to
rewrite the data logic and vice versa.